Fix force push freezing/general issues (#325)
* start fixing force push

* Update force push for progress bar

* commit cuz zion said

* Add a progress bar

* fix globals table

* Add a progress bar again

* Add a progress bar again for table creation

* Add a progress bar again for table creation

* deci update

* update

* update

* update

* update

* remove sleep

* fix globals table

* oops

* fix lock

* clode it

* oops

* bump

---------

+++ binsync/ui/config_dialog.py
@@ -5,6 +5,8 @@ import time
 from pathlib import Path
 from typing import Optional
 
+import filelock
+
 from binsync.core.client import ConnectionWarnings, BINSYNC_ROOT_BRANCH
 from binsync.data.configuration import ProjectConfig
 from binsync.ui.qt_objects import (
@@ -21,7 +23,8 @@ from binsync.ui.qt_objects import (
     QVBoxLayout,
     QTableWidget,
     QTableWidgetItem,
-    QHeaderView
+    QHeaderView,
+    QAbstractItemView
 )
 from binsync.ui.utils import QCollapsibleBox
 
@@ -288,6 +291,8 @@ class ConfigureBSDialog(QDialog):
         self._prev_proj_table.verticalHeader().setVisible(False)
         self._prev_proj_table.horizontalHeader().setVisible(False)
         self._prev_proj_table.setMaximumHeight(50)
+        self._prev_proj_table.setEditTriggers(QAbstractItemView.NoEditTriggers)
+        self._prev_proj_table.itemDoubleClicked.connect(self._handle_prev_proj_double_click)
         prev_proj_layout.addWidget(self._prev_proj_table)
 
         # buttons
@@ -352,6 +357,10 @@ class ConfigureBSDialog(QDialog):
         self.close()
     
     def _on_ok_clicked(self):
+        self.use_recent_project_config()
+        self.close()
+
+    def _handle_prev_proj_double_click(self):
         self.use_recent_project_config()
         self.close()
 
@@ -425,11 +434,10 @@ class ConfigureBSDialog(QDialog):
             push_on_update=not dialog.disable_push, pull_on_update=not dialog.disable_pull,
             commit_on_update=not dialog.disable_commit
         )
-        if successs:
-            self.close()
-        else:
-            self.show()
+        if not successs:
             l.critical("Failed to configure correctly, see above log.")
+
+        self.close()
 
     #
     # Client helpers
@@ -439,12 +447,22 @@ class ConfigureBSDialog(QDialog):
                                   pull_on_update=True, commit_on_update=True):
         lockfile_path = Path(proj_path) / ".git" / "binsync.lock"
         if lockfile_path.exists():
-            box_resp = QMessageBox(self).question(None, "Error", "WARNING: Can only have one binsync client touching a local repository at once." +
-                                                  "If the previous client crashed, the lockfile at:" +
-                                                  f"'{lockfile_path.resolve()}'\n" +
-                                                  "must be deleted. Would you like to delete this now?", QMessageBox.Yes | QMessageBox.No)
-            if box_resp == QMessageBox.Yes:
-                lockfile_path.unlink()
+            repo_lock = filelock.FileLock(lockfile_path)
+            try:
+                repo_lock.acquire(timeout=0)
+                lock_exists = False
+            except filelock.Timeout:
+                lock_exists = True
+
+            if lock_exists:
+                box_resp = QMessageBox(self).question(None, "Error", "WARNING: Can only have one binsync client touching a local repository at once." +
+                                                      "If the previous client crashed, the lockfile at:" +
+                                                      f"'{lockfile_path.resolve()}'\n" +
+                                                      "must be deleted. Would you like to delete this now?", QMessageBox.Yes | QMessageBox.No)
+                if box_resp == QMessageBox.Yes:
+                    lockfile_path.unlink()
+            else:
+                repo_lock.release()
         try:
             connection_warnings = self.controller.connect(
                 username, str(proj_path), init_repo=initialize, remote_url=remote_url,
